Update surface.py
Updated version of surface.py. The method in the Slab class has been modified to allow for the enumerating of different termination sites in a slab. A unittest has been added to see if the .cif files generated to represent the slabs with different termination sites has been properly generated

Former-commit-id: d69f7ad6141dd56651d7a429b9780a28961ddf03 [formerly e31fe220ce42103f3cf2cc968021495ab7e08f17]
Former-commit-id: 7866876b2e227e0b1251182c5c432d28adeca861